+++ modules/os2forms_digital_post/src/Plugin/WebformHandler/WebformHandlerSF1601.php
@@ -3,7 +3,6 @@
 namespace Drupal\os2forms_digital_post\Plugin\WebformHandler;
 
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\os2forms\Plugin\WebformHandler\OS2FormsHandlerInterface;
 use Drupal\os2forms_digital_post\Helper\MeMoHelper;
 use Drupal\os2forms_digital_post\Helper\WebformHelperSF1601;
 use Drupal\webform\Plugin\WebformHandlerBase;
@@ -24,7 +23,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  *   submission = \Drupal\webform\Plugin\WebformHandlerInterface::SUBMISSION_REQUIRED,
  * )
  */
-final class WebformHandlerSF1601 extends WebformHandlerBase implements OS2FormsHandlerInterface {
+final class WebformHandlerSF1601 extends WebformHandlerBase {
   public const MEMO_MESSAGE = 'memo_message';
   public const MEMO_ACTIONS = 'memo_actions';
   public const TYPE = 'type';
@@ -265,37 +264,6 @@ final class WebformHandlerSF1601 extends WebformHandlerBase implements OS2FormsH
     return array_map(static function (array $element) {
       return $element['#title'];
     }, $elements);
-  }
-
-  /**
-   * {@inheritdoc}
-   *
-   * @phpstan-return array<int, mixed>
-   */
-  public function validateConfiguration(): array {
-    $problems = [];
-
-    $recipientElement = $this->configuration[self::MEMO_MESSAGE][static::RECIPIENT_ELEMENT] ?? NULL;
-    if (NULL === $recipientElement || '' === $recipientElement) {
-      $problems[] = $this->t('No recipient element is configured.');
-    }
-    elseif (!isset($this->getRecipientElements()[$recipientElement])) {
-      $problems[] = $this->t('The recipient element (%element) does not exist or does not have a supported type.', [
-        '%element' => $recipientElement,
-      ]);
-    }
-
-    $attachmentElement = $this->configuration[self::MEMO_MESSAGE][static::ATTACHMENT_ELEMENT] ?? NULL;
-    if (NULL === $attachmentElement || '' === $attachmentElement) {
-      $problems[] = $this->t('No attachment element is configured.');
-    }
-    elseif (!isset($this->getAttachmentElements()[$attachmentElement])) {
-      $problems[] = $this->t('The attachment element (%element) does not exist or does not have a supported type.', [
-        '%element' => $attachmentElement,
-      ]);
-    }
-
-    return $problems;
   }
 
   /**